Update config.m
according to email from professor, I don't have to use movie here. But I don't get a better way to implement it. So I keep movie right here. In the loop of challenge, in each iteration I get a completed image, i.e. frame, and I save it in movie. The index i is for the cell movie.